reading code

+++ src/stream/tcb.rs
@@ -173,9 +173,9 @@ impl Tcb {
         if matches!(self.state, TcpState::Established) {
             if let Some(i) = self.inflight_packets.iter().position(|p| p.contains(ack)) {
                 let mut inflight_packet = self.inflight_packets.remove(i);
-                let distance = ack.wrapping_sub(inflight_packet.seq);
-                if (distance as usize) < inflight_packet.payload.len() {
-                    inflight_packet.payload.drain(0..distance as usize);
+                let distance = ack.wrapping_sub(inflight_packet.seq) as usize;
+                if distance < inflight_packet.payload.len() {
+                    inflight_packet.payload.drain(0..distance);
                     inflight_packet.seq = ack;
                     self.inflight_packets.push(inflight_packet);
                 }
